requirements.txt
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
attrs==22.1.0
Babel==2.11.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
cryptography==38.0.1
Django==4.1.2
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-encrypted-model-fields==0.6.5
django-fernet-fields==0.6
docutils==0.19
exceptiongroup==1.0.0
fernet==1.0.1
fontawesome==5.10.1.post1
fontawesomefree==6.2.0
idna==3.4
imagesize==1.4.1
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.4
pyaes==1.6.1
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
requests==2.28.1
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
